annotation: Bump micro version of API jars for compile scope
Since these jars define annotations which are meta-annotated with
Requirement, we want to publish them with the change of maven
dependency to osgi.annotation with compile scope. So we need to bump
the micro version.

This also involves changing *_SPECIFICATION_VERSION constant values to
not include micro versions, so we can append the desired micro version
value in the package-info files.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>